Update event handler setters to travers view and context.
InteractiveView events, like "click," "tap," "hover," etc., can have
function handlers, or be set to a string name. If using a string,
when the handler is set, it attempts to find a matching handler by
traversing the view and the view's context to find a matching
named-handler. The view context is given precedence.